Some bugs have been fixed

+++ repositories/AlbumRepository.js
@@ -60,13 +60,17 @@ const AlbumRepository = {
             }
         })
 
+        if (!album)
+            throw new NotFound("Album not found");
+
+
         await db.MusicAlbum.destroy({
             where: {
                 albumId: id
             }
         })
 
-        
+
 
         await album.addMusics(musics)
